Move TimelineEvents out
